\name{real_data_dsp}
\docType{data}
\alias{real_data_dsp}
\title{A real dataset of genes between the different species.}
\description{
  This data set includes two species and 19330 genes with corresponding short
  read counts, in which the first 143 genes are conserved genes.
}
\usage{real_data_dsp}

\format{A data.frame contains two species and 19330 genes.}
\source{
  Brawand, D., Soumillon, M., Necsulea, A. and Julien, P. et al. (2011).
  The evolution of gene expression levels in mammalian organs. Nature, 478,
  343-348.
}
\keyword{datasets}

