\name{real_data_sp}
\docType{data}
\alias{real_data_sp}
\title{A real dataset of genes between the same species.}
\description{
  This data set includes two samples and each sample with five biological
  replicates and 16519 genes with corresponding short read counts, in which
  the first 530 genes are housekeeping genes.
}
\usage{real_data_sp}

\format{A data.frame contains two samples and each sample with five biological
  replicates and 16519 genes.}
\source{
  Marioni J.C., Mason C.E., et al. (2008). RNA-seq: an assessment of technical    reproducibility and comparisonwith gene expression arrays. Genome Res. 18(9),   1509–1517.
}
\keyword{datasets}

