\name{sim_data_dsp}
\docType{data}
\alias{sim_data_dsp}
\title{A simulation dataset of genes between the differetn species.}
\description{
  This data set includes two species and 18472 genes with corresponding short
  read counts, in which the first 1000 genes are conserved genes.
}
\usage{sim_data_dsp}

\format{A data.frame contains two species and 18472 genes.}
\source{
  Jiadi Zhu, Yan Zhou, Junhui Wang, Youlong Yang (2019, pending publication).
  A minimum enclosing ball method to detect differential expression genes for
  RNA-seq data.
}
\keyword{datasets}

