\name{sim_data_sp}
\docType{data}
\alias{sim_data_sp}
\title{A simulation dataset of genes between the same species.}
\description{
  This data set includes two samples and 10623 genes with corresponding short
  read counts, in which the first 1000 genes are housekeeping genes.
}
\usage{sim_data_sp}

\format{A data.frame contains two samples and 10623 genes.}
\source{
  Jiadi Zhu, Yan Zhou, Junhui Wang, Youlong Yang (2019, pending publication).
  A minimum enclosing ball method to detect differential expression genes for
  RNA-seq data.
}
\keyword{datasets}

