\name{stable_gene_name}
\docType{data}
\alias{stable_gene_name}
\title{The name of the stable genes in all dataset.}
\description{
  The name of the stable genes in all dataset that used to build the model.
}
\usage{stable_gene_name}

\format{A vector contains the name of the stable genes.}
\source{
  The simulation data (sim_scRNA_data) generated from spaltter package.
}
\keyword{datasets}

